+++ spec/scopes_extractor/config_spec.rb
@@ -37,7 +37,7 @@ RSpec.describe ScopesExtractor::Config do
   describe 'http settings' do
     describe '.user_agent' do
       it 'returns the configured user agent' do
-        expect(described_class.user_agent).to include('ScopesExtractor/2.1.2')
+        expect(described_class.user_agent).to include('ScopesExtractor/2.1.3')
         expect(described_class.user_agent).to include('Ruby')
         expect(described_class.user_agent).to include('github.com')
       end